refactor: extract SwapsRescueKey view from Seed.tsx
