+++ views/Swaps/index.tsx
@@ -655,7 +655,7 @@ export default class Swap extends React.PureComponent<SwapProps, SwapState> {
                                     showRescueKeyBtn: true
                                 });
 
-                                navigation.navigate('Seed', {
+                                navigation.navigate('SwapsRescueKey', {
                                     seedPhrase: mnemonic.split(' ')
                                 });
                             }}
@@ -772,7 +772,7 @@ export default class Swap extends React.PureComponent<SwapProps, SwapState> {
                 <TouchableOpacity style={{ marginTop: -10, marginRight: 6 }}>
                     <KeyIcon
                         onPress={() => {
-                            navigation.navigate('Seed', {
+                            navigation.navigate('SwapsRescueKey', {
                                 seedPhrase: this.state.seedPhrase
                             });
                         }}
